Change core features from shell script to Perl
This rewrite implements changes suggested on the Git developer's list [1].

The main reasons for changing the implementation include:
  - Using a combination of 'git diff --raw -z' with 'git update-index /
    checkout-index' improves support for file names with special
    characters.
  - All diff arguments are now passed on to the underlying 'git diff'.
  - File add/deletes are handled more robustly.
  - The script no longer relies on 'mktemp' or 'tar'.

It must be noted that the Perl implementation lost a couple of features
that were in the shell script:
  - The '--copy-back' option is not supported.
  - The names of the tmp directories are now simply 'left' and
    'right' instead of the more descriptive names used before.

These limitations are both the result of the decision to not parse
the diff arguments anymore.  Instead, the arguments are now simply
passed on to the 'git diff --raw' command.  This makes the interface
to 'git diffall' less surprising (i.e. more like 'git diff'), but
makes it very difficult to implement the missing features mentioned
above.

Because the script still relies on 'git-mergetool--lib' to actually
launch the diff tool, it still needs a shell script wrapper to invoke
it.  This is because shell scripts cannot be sourced (or their internal
functions called) directly from Perl.

[1]: http://thread.gmane.org/gmane.comp.version-control.git/191297/focus=191343

Signed-off-by: Tim Henigan <tim.henigan@gmail.com>