Add basic support for submodules
Prior to this commit, if a diff included submodule changes, it would result
in incorrect output shown in the diff tool.  Both sides of the diff would
show an empty directory for the submodule.

This commit teaches diffall to provide basic submodule support. It shows
the change in the submodule SHA1, but does not run an actual diff of the
submodule paths.

Signed-off-by: Tim Henigan <tim.henigan@gmail.com>